refactor cpu as
