+++ src/Assets/Scene.hpp
@@ -6,6 +6,7 @@
 #include <vector>
 #include <glm/vec2.hpp>
 
+#include "CPUAccelerationStructure.h"
 #include "Model.hpp"
 
 namespace Vulkan
@@ -24,12 +25,6 @@ namespace Assets
 	class TextureImage;
 	struct Material;
 	struct LightObject;
-
-	struct AmbientCubeProxy
-	{
-		glm::vec3 position;
-		bool isActive;
-	};
 
 	class Scene final
 	{
@@ -50,7 +45,7 @@ namespace Assets
 			std::vector<AnimationTrack>& tracks);
 		void RebuildMeshBuffer(Vulkan::CommandPool& commandPool,
 			bool supportRayTracing);
-		void RebuildBVH();
+		//void RebuildBVH();
 
 		std::vector<std::shared_ptr<Node>>& Nodes() { return nodes_; }
 		const std::vector<Model>& Models() const { return models_; }
@@ -64,7 +59,6 @@ namespace Assets
 		const Vulkan::Buffer& LightBuffer() const { return *lightBuffer_; }
 		const Vulkan::Buffer& NodeMatrixBuffer() const { return *nodeMatrixBuffer_; }
 		const Vulkan::Buffer& IndirectDrawBuffer() const { return *indirectDrawBuffer_; }
-		const std::vector<AmbientCubeProxy>& AmbientCubeProxys() const { return ambientCubeProxys_; }
 
 		const uint32_t GetLightCount() const {return lightCount_;}
 		const uint32_t GetIndicesCount() const {return indicesCount_;}
@@ -100,8 +94,7 @@ namespace Assets
 
 		void PlayAllTracks();
 
-		Assets::RayCastResult RayCastInCPU(glm::vec3 rayOrigin, glm::vec3 rayDir);
-		void GenerateAmbientCubeCPU();
+		//Assets::RayCastResult RayCastInCPU(glm::vec3 rayOrigin, glm::vec3 rayDir);
 
 		Vulkan::Buffer& AmbientCubeBuffer() const { return *ambientCubeBuffer_; }
 		
@@ -113,7 +106,6 @@ namespace Assets
 		std::vector<LightObject> lights_;
 		std::vector<AnimationTrack> tracks_;
 		std::vector<uvec2> offsets_;
-		std::vector<AmbientCubeProxy> ambientCubeProxys_;
 
 		std::unique_ptr<Vulkan::Buffer> vertexBuffer_;
 		std::unique_ptr<Vulkan::DeviceMemory> vertexBufferMemory_;
@@ -156,5 +148,7 @@ namespace Assets
 		
 		Assets::EnvironmentSetting envSettings_;
 		Camera renderCamera_;
+
+		FCPUAccelerationStructure cpuAccelerationStructure_;
 	};
 }
