+++ src/CMakeLists.txt
@@ -1,8 +1,8 @@
 set_property(GLOBAL PROPERTY USE_FOLDERS ON)
 
 # directory naming
-file(GLOB_RECURSE src_files_assets "Assets/*.cpp" "Assets/*.hpp")
-file(GLOB_RECURSE src_files_utilities "Utilities/*.cpp" "Utilities/*.hpp")
+file(GLOB_RECURSE src_files_assets "Assets/*.cpp" "Assets/*.hpp" "Assets/*.h")
+file(GLOB_RECURSE src_files_utilities "Utilities/*.cpp" "Utilities/*.hpp" "Utilities/*.h")
 file(GLOB_RECURSE src_files_vulkan "Vulkan/*.cpp" "Vulkan/*.hpp")
 file(GLOB_RECURSE src_files_thirdparty 
 	"ThirdParty/json11/*.cpp"
